+++ src/theme/advancedTheme.tsx
@@ -1,0 +1,5 @@
+import { theme } from './theme'
+
+export const advancedTheme = {
+  basicBorder: `${theme.borderWidth} ${theme.borderStyle} ${theme.borderColor}`,
+}
