+++ src/theme/index.tsx
@@ -1,0 +1,6 @@
+import _ from 'lodash/fp'
+
+import { theme as basicTheme } from './theme'
+import { advancedTheme } from './advancedTheme'
+
+export const theme = _.merge(basicTheme, advancedTheme)
