+++ src/theme/theme.tsx
@@ -1,0 +1,25 @@
+export const colors = {
+  black: '0 0 0',
+  grey: '220 220 220',
+}
+
+export const theme = {
+  backgroundColor: '#FFFFFF',
+
+  // Borders
+
+  borderColor: `rgb(${colors.grey})`,
+  borderStyle: 'solid',
+  borderWidth: '1px',
+
+  inputBorderRadius: '20px',
+
+  // Fonts
+  primaryGlobalFont: 'Red Hat Text',
+  secondaryGlobalFont: 'sans-serif',
+  globalFontColor: `rgb(${colors.black})`,
+
+  largeFontSize: '20px',
+  normalFontSize: '12px',
+
+}
