--- src/variables/theme.tsx
@@ -1,12 +1,0 @@
-export const theme = {
-  backgroundColor: '#FFFFFF',
-
-  // Fonts
-  primaryGlobalFont: 'Red Hat Text',
-  secondaryGlobalFont: 'sans-serif',
-  globalFontColor: '0 0 0',
-
-  largeFontSize: '20px',
-  normalFontSize: '12px',
-
-}
